fix: render markdown preview on doc pages
